updated controller for new input data
